requirements.txt
# ============================================================================
# Production Dependencies - AstroGeo AI MLOps Platform
# Tested and verified on Python 3.10+
# ============================================================================

# ============================================================================
# CORE WEB FRAMEWORK
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
email-validator==2.1.0

# ============================================================================
# DATABASE & ORM
# ============================================================================
SQLAlchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
GeoAlchemy2==0.14.2

# ============================================================================
# CACHING
# ============================================================================
redis==5.0.1
hiredis==2.2.3

# ============================================================================
# AUTHENTICATION & SECURITY
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# ============================================================================
# ML & DATA SCIENCE (Core only - no heavy dependencies)
# ============================================================================
numpy==1.26.2
pandas==2.1.3

# Note: Install scikit-learn, tensorflow separately if needed:
# pip install scikit-learn
# pip install tensorflow

# ============================================================================
# AGENTIC AI / LANGCHAIN
# ============================================================================
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1

# ============================================================================
# GEOSPATIAL (Lightweight)
# ============================================================================
shapely==2.0.2
pyproj==3.6.1

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
python-json-logger==2.0.7
structlog==23.2.0

# ============================================================================
# HTTP & NETWORKING
# ============================================================================
httpx==0.25.2
requests==2.31.0

# ============================================================================
# UTILITIES
# ============================================================================
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# ============================================================================
# DATE & TIME
# ============================================================================
python-dateutil==2.8.2
pytz==2023.3.post1

# ============================================================================
# PERFORMANCE
# ============================================================================
orjson==3.9.10